requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5.0
python-dotenv
python-multipart
jinja2
httpx
